pyproject.toml
[project]
name = "CT2STL"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "numpy-stl>=3.2.0",
    "pydicom>=3.0.1",
    "scikit-image>=0.26.0",
    "scipy>=1.17.1",
]

[project.optional-dependencies]
gpu-cuda=["cupy-cuda13x>=14.0.1"]
gpu-rocm=["cupy-rocm-7-0>=14.0.1"]
